linter

+++ test/Atlas.t.sol
@@ -12,7 +12,7 @@ contract AtlasTest is Test {
 
     // Alice's address and private key (EOA with no initial contract code).
     Vm.Wallet alice = vm.createWallet("alice");
-    
+
     // Bob's address and private key (Bob will execute transactions on Alice's behalf).
     Vm.Wallet bob = vm.createWallet("bob");
 
@@ -36,7 +36,11 @@ contract AtlasTest is Test {
         Atlas(alice.addr).addSponsor(bob.addr);
     }
 
-    function getDigest(Atlas.Call[] memory calls, uint256 deadline, uint nonce) internal view returns (bytes32 digest) {
+    function getDigest(Atlas.Call[] memory calls, uint256 deadline, uint256 nonce)
+        internal
+        view
+        returns (bytes32 digest)
+    {
         bytes memory encodedCalls;
         for (uint256 i = 0; i < calls.length; i++) {
             encodedCalls = abi.encodePacked(encodedCalls, calls[i].to, calls[i].value, calls[i].data);
@@ -44,7 +48,11 @@ contract AtlasTest is Test {
 
         // IMPORTANT!! `Atlas(alice.addr).DOMAIN_SEPARATOR()` need ot be called from alice bytecodes because it doesn't have the same address as the atlas deployed one.
         digest = keccak256(
-            abi.encodePacked(hex"1901", Atlas(alice.addr).DOMAIN_SEPARATOR(), keccak256(abi.encodePacked(deadline, nonce, encodedCalls)))
+            abi.encodePacked(
+                hex"1901",
+                Atlas(alice.addr).DOMAIN_SEPARATOR(),
+                keccak256(abi.encodePacked(deadline, nonce, encodedCalls))
+            )
         );
     }
 
@@ -84,7 +92,7 @@ contract AtlasTest is Test {
         r = bytes32(0);
 
         vm.expectRevert(); // Expect to revert because of the wrong value in the call (invalid signature)
-        
+
         vm.broadcast(bob.privateKey);
         Atlas(alice.addr).execute(calls, deadline, v, r, s);
     }
@@ -107,7 +115,7 @@ contract AtlasTest is Test {
         Atlas(alice.addr).execute(calls, deadline, v, r, s);
 
         vm.expectRevert(); // we should not be able to resend the same call signed
-        
+
         vm.broadcast(bob.privateKey);
         Atlas(alice.addr).execute(calls, deadline, v, r, s);
     }
